Split remaining test monoliths
